% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{.onLoad}
\alias{.onLoad}
\title{Set graphic defaults}
\usage{
.onLoad(libname, pkgname)
}
\value{
None
}
\description{
Change the default font to Avenir Next, and the default graphical element color to green
}
\concept{addlogo, theme_citylab theme_citylab_fb theme_citylab_map citylab_colors}
