% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dhm.R
\name{theme_dhm}
\alias{theme_dhm}
\title{David H. Montgomery's ggplot theme}
\usage{
theme_dhm(base_size = 10, base_family = "Avenir Next")
}
\arguments{
\item{base_size}{Sets the base font size for the plot. Default is 12.}

\item{base_family}{Sets the base font for the plot. Default is Avenir Next.}
}
\value{
None
}
\description{
Applies David's graphical styles to a ggplot object.
}
\examples{
\dontrun{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_dhm() + labs(title = "Title", caption = "Caption")}

}
\concept{theme_reset defaults.R}
